pyproject.toml
[project]
name = "devbot"
version = "0.1.0"
description = "Autonomous DevOps Agent"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.18.0",
    "watchdog>=4.0.0",
    "GitPython>=3.1.41",
    "PyGithub>=2.1.1",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-dotenv>=1.0.1",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "httpx>=0.26.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "mypy>=1.8.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
strict = true
